requirements.txt
flask
flask-cors
joblib
numpy
scikit-learn
mysql-connector-python
pandas
scipy
gunicorn